Tests/LibWeb/Text/expected/abspos-relayout-index-compaction.txt

layout-passes-sum-matches=true
full-passes-bounded=false
boundary=40,30,200,40
FAIL
